Hi, I’m Andrew, can you point me in the right direction?
I recently graduated as a Life Coach and have the idea for an app which will use music to help some of my clients to manage their emotional state.  The idea is for the client to set the music with a beat which matches their emotion, and then, after a period of synchronisation, the app will gradually reduce the beat over time to help them to relax, or will gradually increase the beat to help them energise themselves when they are feeling down.  

As the beat will need to be set at the beginning and then progressively change over time, I cannot think of any alternative but to use Computer Music.  This would also give the benefit of being able to choose different genres of music to match people’s personal preferences and moods. 

The problem is that I do not really know anything about Computer Music and so would like to ask if Magenta or GitHub will allow me to achieve the effect I am looking for.  If it is not the right solution, I would be very grateful if you could point me in the right direction to get the help I need.

I really would be grateful for any advice you can give me.

Thank you in advance,

Andrew Simpson